revert condtion on parameter update in test_default_db. move televir zip to views, create view for download
