+++ pathogen_identification/views.py
@@ -1168,3 +1168,49 @@ def download_file_ref(requestdst):
             ] = "attachment; filename=%s" % os.path.basename(filepath)
             # Return the response value
             return response
+
+
+
+import zipfile
+
+def generate_zip_file(file_list: list, zip_file_path: str) -> str:
+        
+        with zipfile.ZipFile(zip_file_path, "w") as zip_file:
+            for file_path in file_list:
+                zip_file.write(file_path, os.path.basename(file_path))
+    
+        return zip_file_path
+
+def get_create_zip(file_list: list, outdir: str, zip_file_name: str) -> str:
+
+    zip_file_path = os.path.join(outdir, zip_file_name)
+
+    if os.path.exists(zip_file_path):
+        os.unlink(zip_file_path)
+
+    zip_file_path = generate_zip_file(file_list, zip_file_path)
+
+    return zip_file_path
+
+def download_intermediate_report_zipfile(request):
+    """
+    download intermediate report files in zip"""
+
+    if request.method == "POST":
+
+        run_pk= request.GET.get("run_pk")
+        run_main= RunMain.objects.get(pk= int(run_pk))
+
+        intermediate_reports= run_main.intermediate_reports_get()
+        run_main_dir= infer_run_media_dir(run_main)
+        zip_file_name= "{}_intermediate_reports.zip".format(run_main.name)
+
+        zip_file_path= get_create_zip(intermediate_reports, run_main.static_dir, zip_file_name)
+
+        path= open(zip_file_path, "rb")
+        mime_type, _= mimetypes.guess_type(zip_file_path)
+        response= HttpResponse(path, content_type= mime_type)
+        response["Content-Disposition"]= "attachment; filename={}".format(zip_file_name)
+
+        return response
+
